Fix problem and Update Implement
Fix:
If there are two player and one player draws three cards and bust, the game will regard there is no player draws card.
So I set up a if to check when the player list have player bust, it will return the player not bust.
Implement: Points Level 1 and first and second level 2